requirements.txt

Flask-Bootstrap==3.3.7.1 
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3


Flask-SQLAlchemy==2.4.4

SQLAlchemy==1.3.19
PyJWT==2.0.0
Flask-WTF==0.14.3
alembic==1.0.0
asn1crypto==0.24.0
blinker==1.6.2
cffi==1.15.1
click==8.1.6
cryptography==41.0.3
dominate==2.8.0
favicon==0.7.0
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Moment==1.0.5
flask-paginate==2022.1.8
Flask-Script==2.0.6
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
mako==1.2.4
MarkupSafe==2.1.3
Pillow==10.0.0
pycparser==2.21
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.19
visitor==0.1.3
Werkzeug==2.3.6
WTForms==3.0.1